# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TaskTrial APIs",
    "version": "1.0.0",
    "description": "Project Management App",
    "contact": {
      "email": "tasktrial27@gmail.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/departments/all": {
      "get": {
        "tags": ["Department"],
        "summary": "Get all departments",
        "description": "Retrieve a paginated list of all departments.",
        "operationId": "getAllDepartments",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Departments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllDepartmentsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
    "/api/departments/organization/{organizationId}/manager/{managerId}": {
      "post": {
        "tags": ["Department"],
        "summary": "Create a new department",
        "description": "Allows authorized users to create a new department within a specific organization and assign a manager.",
        "operationId": "createDepartment",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "description": "ID of the organization where the department will be created",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "managerId",
            "in": "path",
            "required": true,
            "description": "ID of the user to be assigned as the department manager",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Details of the department to be created",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Department created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDepartmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - department with the same name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/departments/{departmentId}/": {
      "put": {
        "tags": ["Department"],
        "summary": "Create a new department",
        "description": "Update the details of a specific department, including its name and description.",
        "operationId": "updateDepartment",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "departmentId",
            "in": "path",
            "required": true,
            "description": "ID of the department to update",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Department update data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentRequest"
              },
              "examples": {
                "updateNameAndDescription": {
                  "value": {
                    "name": "Updated Department Name",
                    "description": "Updated description for the department"
                  }
                },
                "updateNameOnly": {
                  "value": {
                    "name": "New Department Name"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateDepartmentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "validationError": {
                    "value": {
                      "success": false,
                      "message": "Validation failed",
                      "errors": ["\"name\" must be at least 2 characters long"]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Department not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "notFound": {
                    "value": {
                      "success": false,
                      "message": "Department not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/departments/{id}": {
      "get": {
        "tags": ["Department"],
        "summary": "Get department details",
        "description": "Retrieve detailed information about a specific department.",
        "operationId": "getDepartmentById",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the department to retrieve",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Department details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDepartmentResponse"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      },
      "put": {
        "tags": ["Department"],
        "summary": "Update department details",
        "description": "Update the details of a specific department.",
        "operationId": "updateDepartment",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the department to update",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateDepartmentResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequestError" },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      },
      "delete": {
        "tags": ["Department"],
        "summary": "Soft delete a department",
        "description": "Mark a department as deleted.",
        "operationId": "softDeleteDepartment",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the department to delete",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Department soft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoftDeleteDepartmentResponse"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      }
    },
    "/api/departments/{id}/restore": {
      "patch": {
        "tags": ["Department"],
        "summary": "Restore a soft-deleted department",
        "description": "Restore a department that was previously soft-deleted.",
        "operationId": "restoreDepartment",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the department to restore",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Department restored successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreDepartmentResponse"
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFoundError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Validation error details"]
          }
        }
      },
...
      "CreateDepartmentRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Engineering" },
          "description": {
            "type": "string",
            "example": "Handles engineering tasks"
          },
          "organizationId": { "type": "string", "format": "uuid" },
          "managerId": { "type": "string", "format": "uuid" }
        },
        "required": ["name", "organizationId", "managerId"]
      },
      "CreateDepartmentResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Department created successfully"
          },
          "department": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "name": { "type": "string" },
              "description": { "type": "string" },
              "organizationId": { "type": "string", "format": "uuid" },
              "managerId": { "type": "string", "format": "uuid" }
            }
          }
        }
      },
      "GetAllDepartmentsResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Departments retrieved successfully"
          },
          "departments": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "format": "uuid" },
                "name": { "type": "string" },
                "description": { "type": "string" },
                "organization": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "name": { "type": "string" }
                  }
                },
                "manager": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "firstName": { "type": "string" },
                    "lastName": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "GetDepartmentResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Department retrieved successfully"
          },
          "department": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "name": { "type": "string" },
              "description": { "type": "string" },
              "organization": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "format": "uuid" },
                  "name": { "type": "string" }
                }
              },
              "manager": {
                "type": "object",
                "properties": {
                  "id": { "type": "string", "format": "uuid" },
                  "firstName": { "type": "string" },
                  "lastName": { "type": "string" }
                }
              }
            }
          }
        }
      },
      "UpdateDepartmentRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "Updated Department Name" },
          "description": { "type": "string", "example": "Updated description" }
        }
      },
      "UpdateDepartmentResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Department updated successfully"
          },
          "department": {
            "type": "object",
            "properties": {
              "id": { "type": "string", "format": "uuid" },
              "name": { "type": "string" },
              "description": { "type": "string" }
            }
          }
        }
      },
      "SoftDeleteDepartmentResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Department soft deleted successfully"
          }
        }
      },
      "RestoreDepartmentResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Department restored successfully"
          }
        }
      }
    }
  }
}